refactor per discussion between S. Henshaw and D. Slater on 11 May 2022
