+++ armory/datasets/builder/Dockerfile
@@ -10,8 +10,8 @@
 FROM tensorflow/tensorflow
 
 WORKDIR dataset_builder
-COPY ./requirements.txt /dataset_builder/
+COPY requirements.txt /dataset_builder/
 RUN pip install -r /dataset_builder/requirements.txt
 
 COPY ./*.py /dataset_builder/
-COPY ./build_classes /dataset_builder/build_classes/
+COPY build_classes /dataset_builder/build_classes/
