+++ armory/datasets/builder/build_classes/adversarial/apricot_metadata.py
