+++ armory/datasets/builder/build_classes/digit.py
