+++ armory/datasets/builder/build_classes/german_traffic_sign.py
