+++ armory/datasets/builder/build_classes/ucf101_clean.py
