+++ tests/datasets/resources/example_built_datasets/my_dataset/1.0.0/my_dataset-train.tfrecord-00000-of-00001
