+++ tests/datasets/resources/example_local_class_directory/__init__.py
