+++ tests/datasets/resources/example_local_class_directory/checksums.tsv
