+++ tests/datasets/resources/example_local_class_directory/dummy_data/train_imgs/I00000.jpg
