+++ tests/datasets/resources/example_local_class_directory/my_dataset.py
