--- tools/dataset_builder/tests/__init__.py
@@ -1,1 +1,0 @@
-# THIS NEEDS TO BE HERE FOR TESTS TO WORK
